requirements.txt
tensorflow-gpu
keras
scikit-learn
hyperas
pandas
deepdish
matplotlib
ipykernel
